requirements.txt
flask==2.3.3
pandas==2.0.3
matplotlib==3.7.2
numpy==1.26.4
reportlab
mysql-connector-python==8.0.33